refactor: move response utils to submission module
