+++ src/app/modules/submission/submission.types.ts
@@ -1,0 +1,6 @@
+import { FieldResponse } from 'src/types'
+
+export type ProcessedFieldResponse = FieldResponse & {
+  isVisible?: boolean
+  isUserVerified?: boolean
+}
